AGENT: added method to agent to initialise AIlogic by setting up a new state machine and assigning a seeks state
added a method to convert XMFOAT3 to Vector2f

added a method that sets the agents velocity

added an Update method so that the agent can process instructions per frame as this was missing

LEVEL1
added an Agent called m_enemy that is updated in the update method just like the cude object.

AILOGIC
added an AIStateMachine, an AIState and an AISeek so that an agent invokes its own statemachine that processes the ai states that have been passed to it so that the agent becomes autonomous with behaviuors.