Added an event emitter to the input class that sends an event when the left mouse button is clicked.
This is to be used as a temporary mehod of getting a position in the game for an effect to occur. Which will be used to test and debug a projectile's starting position.

The next step is to create a class that listens for this event and places a projectile at the position that the mouse was clicked.