refactored player class's position vector to use smart pointers instead of raw pointers
added an off set and a maxlifetime variable to the projectile class. setting the offset allows the projectile to spawn at a location in relation to a passed in vector. Having a max lifetime variable allows the projectile to remember the initial lifetime instead of being over written each frame.

The projectile editor now records a spawn position for each manager's pool of projectiles.
the number of projectiles in a manager is now shown and can be increased by clicking the "Add Projectile" button.
The TestButtons method now correctly refills the list of managers built from the loaded json file. This includes the projectile pool for each manager and begins the projectile animation.

removed linkage to events in the destructor of the projectile manager class as this was causing errors because there where events without clients in the eventsystem.
Removed duplicated and unused SpawnProjectile methods while also refactoring a method to receive a spawn location with an optional life time parameter.