stubbing out the process of getting a setting from the projectile editor that enables a projectile manager to repeat its pattern of emitting projectiles.
by passing a flag from the editor, the manager will count the delays and lifetimes of each projectile for a total duration. using this and the received flag, the manager will reset a counter that tracks the total duration. when the counter reaches zero AND is meant to repeat it's cycle, the counter is reset to the total duration.

the next step is to call the method that spawns all projectiles for the individual manager.