updated the player controller to use const smart references. Previously, raw pointers and mutating components, reducing performance and causing issues trying to reference said components.
implemented an offset vector2f in the player controller, because the spawn position being passed results in projectiles being placed at 0,0 of the sprite.
Now, the spawn position of the sprite uses an of set calculated from the sprite height and width to place the projectile at the center of the sprite.

added a player shooting class to capture user inputs to setup and execute emitter functions.
Previously, nothing accomplished this.

update the emitter class to use const mart reference to the projectile manager to stop mutation and increase performance.